restructure project to separate source files / includes files in prep for there being another binary in the project
