+++ CMakeLists.txt
@@ -6,23 +6,25 @@ set(CMAKE_CXX_STANDARD 17)
 
 add_executable(
         akaiutil
-        akaiutil.c
-        akaiutil.h
-        akaiutil_file.c
-        akaiutil_file.h
-        akaiutil_io.c
-        akaiutil_io.h
-        akaiutil_main.c
-        akaiutil_take.c
-        akaiutil_take.h
-        akaiutil_tar.c
-        akaiutil_tar.h
-        akaiutil_wav.c
-        akaiutil_wav.h
-        commoninclude.h
-        commonlib.c
-        fdrawcmd.h
+        akaiutil/akaiutil.c
+        include/akaiutil/akaiutil.h
+        akaiutil/akaiutil_file.c
+        include/akaiutil/akaiutil_file.h
+        akaiutil/akaiutil_io.c
+        include/akaiutil/akaiutil_io.h
+        akaiutil/akaiutil_main.c
+        akaiutil/akaiutil_take.c
+        include/akaiutil/akaiutil_take.h
+        akaiutil/akaiutil_tar.c
+        include/akaiutil/akaiutil_tar.h
+        akaiutil/akaiutil_wav.c
+        include/akaiutil/akaiutil_wav.h
+        include/akaiutil/commoninclude.h
+        akaiutil/commonlib.c
+        include/akaiutil/fdrawcmd.h
 )
+
+target_include_directories(akaiutil PRIVATE include)
 
 # For now, link libm on UNIX
 # TODO(tdial): check on mac
