+++ akaiutil/akaiutil_file.c
@@ -18,10 +18,10 @@
 * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 */
 
-#include "commoninclude.h"
-#include "akaiutil.h"
-#include "akaiutil_file.h"
-#include "akaiutil_wav.h"
+#include "akaiutil/commoninclude.h"
+#include "akaiutil/akaiutil.h"
+#include "akaiutil/akaiutil_file.h"
+#include "akaiutil/akaiutil_wav.h"
 
 /* AKAI file info */
 int akai_file_info(struct file_s *fp, int verbose) {
