+++ akaiutil/commonlib.c
@@ -18,11 +18,7 @@
 * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 */
 
-
-
-#include "commoninclude.h"
-
-
+#include "akaiutil/commoninclude.h"
 
 #ifdef _VISUALCPP
